Refactor code after peer review
Moved the environment variables out of the controller and put them
into an initialiser so is a single point of definition and setting

There were discussions about trying to change the vars names to have
a prefix of "SETTINGS__" like our other environment var but we cannot
do that because Platforms team relies on the naming of these variables
for their scripts and the values are set as part of a jenkins job.

By adding them to Application config variable it allowed us to improve
the tests by not having to change env variables as part of the tests.