Switched to using Settings instead of initializers
The values we need to pull in from the Docker container are
picked up in run.sh. We use them to set our own system environment
variables (i.e prefixed with 'SETTINGS_').

Default settings for instances where the environment variables
are not set are declared in config/settings.yml

The reason it was done this way is to ensure all environment
variables are used in a consistent way across the app. It also
helps test that the endpoints return the expected values.