CT-508 Improve GOVUK Search API call
This implements a basic snowball stop list of words to
filter to. It takes the original topic and removes the stop words.

We are using a gem called stoplist-filter which has its own list of
stops words that can be extended. It also has locales support should
the service need to be translated into other languages.