CT-605 Authenticate incoming correspondence items
This PR adds an extra step to the process of a member of the public submitting
a correspondence item.  When the item is created, a mail with a URL comprising the
item's uuid is sent to the specified email address, and the item is not consiered
authenticated until the user has clicked on the link in the email.