CT-753 Fix bug Duplicate mails being sent
This bug is caused by us enquueing a mail-sender job on the sidekiq queue, and then both
serviers picking up the job and executing it.

This is solved by converting the code to use the ActionMailer deliver_later method which
uses ActiveJob behind the scenes and takes care of concurrency issues like this.