CT-4199 Remove feedback links from accessibility page
This is an attempt to try to reduce spam received through the feedback form. We are removing the links from the accessibility page as this form doesn't belong here, but we are leaving the link in the remaining places.

Renamed the URL (as the spam bot might reuse the old URL).

Also fixed a minor typo I found in the copy.

Note: we might need to iterate on this if we find out spam is still received, but for now this is a first step.