Bump Rails and dependencies
Bump Rails to the latest patch in the 6.1.x branch, along with several other dependencies to their minor bugfix versions.

This resolves a bunch of dependabot alerts.

The remaining ones, if any, will be treated in a case by case basis as it may not be possible to bump to the latest version straight away (for example `Sprockets 4.x`)

Also, fixed deprecation warning with `Capybara::Selenium` options.