Bump i18n-tasks gem and a few others (#599)
* Bump i18n-tasks gem and a few others

This is also safe to bump. No breaking changes and keeps working as intended.

Removed `rack` explicit gem as it is a dependency of ActionPack which gets installed as part of Rails anyways.

Also, resolved another deprecation warning in `pry-byebug`.

* Disable dependabot version updates

Update `dependabot.yml` to only raise PR for security versions, and not for version updates, as per documentation here:

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#open-pull-requests-limit

This is intended to reduce the amount of noisy PRs we get, so we can bump the gem versions (if convenient) at our own pace.